DataChunk::size is now of type size_t

+++ include/ProtoZed/DataChunk.h
@@ -29,7 +29,7 @@ namespace PZ
 	public:
 		DataChunk() : data(nullptr), size(0), references(new unsigned int(1))
 		{}
-		DataChunk(const char *data, unsigned int size) : data(data), size(size), references(new unsigned int(1))
+		DataChunk(const char *data, size_t size) : data(data), size(size), references(new unsigned int(1))
 		{}
 		DataChunk(const DataChunk &other) : data(other.data), size(other.size), references(other.references)
 		{
@@ -61,14 +61,14 @@ namespace PZ
 			return data;
 		}
 
-		unsigned int GetSize() const
+		size_t GetSize() const
 		{
 			return size;
 		}
 
 	private:
 		const char *data;
-		unsigned int size;
+		size_t size;
 
 		void addReference()
 		{
